requirements.txt
discord.py
python-dotenv
yt-dlp
PyNaCl
pomice>=2.10.0
